requirements.txt
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
